ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Core ML/DL
torch==2.1.2
torchvision==0.16.2
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.24.3
pillow==10.1.0

# OCR
pytesseract==0.3.10
pdf2image==1.17.1

# Utilities
pydantic==2.5.0
requests==2.31.0
